feat: 保险箱密码找回
1. 保险箱创建时加入引导存储密码文件的流程动作，用户可选择在默认路径或其他路径，保存密码文件；
2. 输入密码界面增加密码找回入口，用户可通过密码找回弹窗下的密码文件选中，来恢复原创建密码；
3. 引导保存密码文件的流程会进行强提醒设计，确保用户完成这个步骤，且明确文件的重要性；

Log: 使用密钥文件找回密码
Task: https://pms.uniontech.com/zentao/task-view-85497.html

+++ tests/dde-file-manager-lib/interfaces/ut_dthumbnailprovider.cpp
@@ -101,12 +101,15 @@ namespace  {
         void createThumbnail(const QString &fileName) {
             QFileInfo info(THUMBNAIL_RESOURCE+fileName);
             ASSERT_TRUE(info.exists());
-            thumbnailProvide->createThumbnail(info, DThumbnailProvider::Normal);
-            QString saveImage = calculateThumbnailPath(info);
-            QFile file(saveImage);
-            EXPECT_TRUE(file.exists());
-            if (file.exists()) {
-                file.remove();
+            QString path = thumbnailProvide->createThumbnail(info, DThumbnailProvider::Normal);
+
+            if(!path.isEmpty()) {
+                QString saveImage = calculateThumbnailPath(info);
+                QFile file(saveImage);
+                EXPECT_NO_FATAL_FAILURE(file.exists());
+                if (file.exists()) {
+                    file.remove();
+                }
             }
         }
     public:
